Projects/Real_Time_Weather/requirements.txt
requests
python-dotenv
schedule
streamlit